+++ simulation_online/include/softmax_layer.hpp
@@ -9,11 +9,11 @@ void softmax_layer_backward_float(const float* softmax_outputs,
                             float* softmax_gradients, 
                             const uint8_t softmax_batch_size, 
                             const uint8_t softmax_num_labels);
-void softmax_layer_fixed(const int32_t* softmax_inputs, int32_t* softmax_outputs, const uint8_t softmax_batch_size, const uint8_t softmax_num_labels, const float rescale_value, const uint8_t activation_bits);
+void softmax_layer_fixed(const int32_t* softmax_inputs, int32_t* softmax_outputs, const uint8_t softmax_batch_size, const uint8_t softmax_num_labels, const double rescale_value, const uint8_t activation_bits);
 void softmax_layer_backward_fixed(const int32_t* softmax_outputs, 
                             const int32_t* true_labels, 
                             int32_t* softmax_gradients, 
                             const uint8_t softmax_batch_size, 
                             const uint8_t softmax_num_labels,
-                            const float rescale_value);
+                            const double rescale_value);
 #endif
