+++ simulation_online/lib/avgpool2d_layer.cpp
@@ -60,7 +60,7 @@ void avgpool2d_layer_float(float *avgpool2d_inputs, float *avgpool2d_outputs,
 void avgpool2d_layer_fixed(int32_t *avgpool2d_inputs, int32_t *avgpool2d_outputs,
     const uint16_t avgpool2d_input_width, const uint16_t avgpool2d_input_height, const uint16_t avgpool2d_input_depth,
     const uint8_t avgpool2d_stride, const uint8_t avgpool2d_kernel_width, const uint8_t avgpool2d_kernel_height, const uint8_t avgpool2d_batch_size,
-    const uint8_t avgpool2d_pad_top, const uint8_t avgpool2d_pad_bottom, const uint8_t avgpool2d_pad_left, const uint8_t avgpool2d_pad_right, const float rescale_value, const uint8_t activation_bits) {
+    const uint8_t avgpool2d_pad_top, const uint8_t avgpool2d_pad_bottom, const uint8_t avgpool2d_pad_left, const uint8_t avgpool2d_pad_right, const double rescale_value, const uint8_t activation_bits) {
 
     uint32_t avgpool2d_batch_input_offset = 0;
     uint32_t avgpool2d_batch_output_offset = 0;
