+++ simulation_online/lib/dw_conv_layer.cpp
@@ -77,7 +77,7 @@ void dw_conv_layer_float(float *dw_input_features, float *dw_output_features, fl
 void dw_conv_layer_fixed(int32_t *dw_input_features, int32_t *dw_output_features, int32_t *dw_kernel_weights, int32_t *dw_kernel_biases,
     const uint16_t dw_input_width, const uint16_t dw_input_height, const uint16_t dw_input_depth,
     const uint8_t dw_stride, const uint8_t dw_kernel_width, const uint8_t dw_kernel_height, const uint8_t dw_batch_size,
-    const uint8_t dw_pad_top, const uint8_t dw_pad_bottom, const uint8_t dw_pad_left, const uint8_t dw_pad_right, const float rescale_value, const uint8_t activation_bits) {
+    const uint8_t dw_pad_top, const uint8_t dw_pad_bottom, const uint8_t dw_pad_left, const uint8_t dw_pad_right, const double rescale_value, const uint8_t activation_bits) {
     uint32_t dw_batch_input_offset = 0;
     uint32_t dw_batch_output_offset = 0;
